- Example L04 now uses D3D9 -Color is now a struct -Vector3Di/f/d is now a struct -Vertex3D(TTCoords/Tangents) is now a struct These structs can be casted to their native type in most cases. Some have a ToNative() function. Every class and example has been updated to use them correctly, but not all of them have been tested.

+++ examples/05.UserInterface/Program.cs
@@ -88,7 +88,7 @@ namespace _05.UserInterface
 							foreach (GUIDefaultColor which in Enum.GetValues(typeof(GUIDefaultColor)))
 							{
 								Color color = env.Skin.GetColor(which);
-								color.Alpha = pos;
+								color.Alpha = (uint)pos;
 								env.Skin.SetColor(color, which);
 							}
 						}
